Move elfObjDisasm so it can be used with the spimdisasm pip installation
