+++ spimdisasm/elfObjDisasm/__init__.py
@@ -1,0 +1,9 @@
+#!/usr/bin/env python3
+
+# SPDX-FileCopyrightText: © 2022 Decompollaborate
+# SPDX-License-Identifier: MIT
+
+from __future__ import annotations
+
+
+from .ElfObjDisasmInternals import getArgsParser, applyArgs, applyGlobalConfigurations, getOutputPath, getProcessedSections, changeGlobalSegmentRanges, insertSymbolsIntoContext, injectAllElfSymbols, processGlobalOffsetTable, elfObjDisasmMain
